requirements.txt
absl-py==2.2.2
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Authlib==1.6.5
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==6.2.1
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
Django==5.2
docstring_parser==0.17.0
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.121.1
fastjsonschema==2.19.0
Flask==3.0.2
flatbuffers==25.2.10
fonttools==4.46.0
fqdn==1.5.1
gast==0.6.0
google-adk==1.18.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.126.1
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.5.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.49.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
ibm-cloud-sdk-core==3.19.2
ibm-watson==8.0.0
idna==3.6
importlib_metadata==8.7.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.9.2
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mcp==1.21.0
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.0.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
mysql-connector-python==8.3.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.15.0
overrides==7.4.0
packaging==25.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pbr==6.0.0
Pillow==10.1.0
platformdirs==4.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
proto-plus==1.26.1
protobuf==6.33.0
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.10.1
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
pywin32==311
pywinpty==2.0.12
PyYAML==6.0.3
pyzmq==25.1.1
qiskit==1.0.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.13.2
rsa==4.9.1
rustworkx==0.14.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==69.0.2
shapely==2.1.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.5.1
soupsieve==2.5
SpeechRecognition==3.14.4
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
stevedore==5.2.0
symengine==0.11.0
sympy==1.12
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.1.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==15.0.1
Werkzeug==3.0.1
wheel==0.45.1
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.17.2
zipp==3.23.0
